requirements.txt
# web app
flask==0.11.1
python-dotenv
cf-deployment-tracker
cloudant
couchdbkit
gensim
couchdb
nltk